!2457 feat(#I4V64R): TotalCount parameter on Pagination/ListView change long to int type * refactor: 更改参数类型为 Int

+++ src/BootstrapBlazor/Components/ListView/ListView.razor.cs
@@ -81,7 +81,7 @@ public partial class ListView<TItem> : BootstrapComponentBase where TItem : clas
     /// <summary>
     /// 获得/设置 数据总条目
     /// </summary>
-    protected long TotalCount { get; set; }
+    protected int TotalCount { get; set; }
 
     /// <summary>
     /// 获得/设置 当前页码
